refactor etl file structure (#110)
refactor etl file structure
