+++ app/etl/organisation/comtrade.py
@@ -1,4 +1,4 @@
-from app.etl.etl_snapshot_data import SnapshotDataPipeline
+from app.etl.pipeline_type.snapshot_data import SnapshotDataPipeline
 
 
 class ComtradeCountryCodeAndISOPipeline(SnapshotDataPipeline):
