+++ app/etl/organisation/ons.py
@@ -1,6 +1,6 @@
 from io import BytesIO
 
-from app.etl.etl_snapshot_data import SnapshotDataPipeline
+from app.etl.pipeline_type.snapshot_data import SnapshotDataPipeline
 
 
 class ONSPostcodeDirectoryPipeline(SnapshotDataPipeline):
