+++ tests/etl/test_dit_eu_country_membership.py
@@ -1,6 +1,6 @@
 from datatools.io.fileinfo import FileInfo
 
-from app.etl.etl_dit_eu_country_membership import DITEUCountryMembershipPipeline
+from app.etl.organisation.dit import DITEUCountryMembershipPipeline
 from tests.utils import rows_equal_table
